Move single instance handling to a separate static class and to the beginning of Main (misc settings split between store and presentation)
